requirements.txt
requests~=2.32.0
python-dotenv~=1.1.0
pydantic~=2.11.0
mysql-connector-python~=9.4.0